newBack/requirements.txt
alembic==1.8.1
anyio==3.6.2
bcrypt==4.0.1
fastapi==0.87.0
FastAPI-SQLAlchemy==0.2.1
GeoAlchemy==0.7.2
greenlet==2.0.1
idna==3.4
inflect==6.0.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pydantic==1.10.2
PyPDF2==2.11.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
reportlab==3.6.12
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
typing_extensions==4.4.0
